+++ test/test_scatter3D.py
@@ -19,7 +19,7 @@ def test_scatter3d():
         ]
         for _ in range(80)
     ]
-    scatter3d = Scatter3D("3D 散点图示例", width=1200, height=600)
+    scatter3d = Scatter3D("3D 散点图示例", width='1200px', height='600px')
     scatter3d.add("", data, is_visualmap=True, visual_range_color=RANGE_COLOR)
     scatter3d.render()